added hook for the favourite
